Version 4 and 5 PWE computeSalary Changes return of computeSalary method change from float to double

+++ EmployeeProject/src/employee/version4/PieceWorkerEmployee.java
@@ -37,7 +37,7 @@ public class PieceWorkerEmployee extends Employee {
         this.ratePerPiece = ratePerPiece > 0 ? ratePerPiece : 0;
     }
 
-    public float computeSalary(){
+    public double computeSalary(){
         int bonus = totalPiecesFinished / 100;
         return (totalPiecesFinished * ratePerPiece) + (bonus * (ratePerPiece * 10));
     }
